+++ Sources/Action/Action.swift
@@ -2,29 +2,4 @@
 
 import Foundation
 
-public final class AnyAction: Decodable {
-    public static var builder: ActionBuilderProtocol?
-
-    private enum CodingKeys: CodingKey {
-        case identifier
-    }
-
-    public let action: Action
-
-    public init(action: Action) {
-        self.action = action
-    }
-
-    public required init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        let identifier = try container.decode(String.self, forKey: .identifier)
-        guard let action = Self.builder?.action(from: decoder, withIdentifier: identifier) else {
-            throw TaksiError.decodingError
-        }
-        self.action = action
-    }
-}
-
 public protocol Action: AnyObject {}
-
-public protocol DecodableActionProtocol: Action, Decodable {}
