+++ Sources/Action/ActionType/ActionType.swift
@@ -3,7 +3,3 @@
 import Foundation
 
 protocol ActionType: Decodable {}
-
-protocol VisitableActionType {
-    func visit(from visitor: ActionTypeVisitor, using decoder: Decoder) -> Action?
-}
