+++ Sources/Action/ActionType/VisitableActionType.swift
@@ -1,0 +1,7 @@
+// Created by Mateus Lino
+
+import Foundation
+
+protocol VisitableActionType {
+    func visit(from visitor: ActionTypeVisitor, using decoder: Decoder) -> Action?
+}
