+++ Sources/Action/DecodableAction.swift
@@ -1,0 +1,5 @@
+// Created by Mateus Lino
+
+import Foundation
+
+public protocol DecodableAction: Action, Decodable {}
