+++ Sources/Interface/Component/TaksiComponentIdentifier.swift
@@ -5,7 +5,7 @@ import Foundation
 enum TaksiComponentIdentifier: String {
     case collection = "collection_component"
 
-    var metatype: any DecodableComponentProtocol.Type {
+    var metatype: any DecodableComponent.Type {
         switch self {
         case .collection:
             return CollectionComponent.self
