+++ Sources/UI/ComponentViewRenderMode.swift
@@ -1,0 +1,10 @@
+// Created by Mateus Lino
+
+import SwiftUI
+
+public typealias ComponentViewRenderModeBinding = Binding<ComponentViewRenderMode>
+
+public enum ComponentViewRenderMode {
+    case skeleton
+    case content
+}
