+++ Sources/UI/ComponentViews/EmptyComponentView.swift
@@ -1,0 +1,5 @@
+// Created by Mateus Lino
+
+import Foundation
+
+public struct EmptyComponentView: ComponentView {}
